Add test suite support to opendeplete
This addition adds a mostly complete test suite for opendeplete.  It
includes unit tests for the more easily unit-testable portions of the
code, integral tests for the integrator routines, and a full system
test to ensure that all components work in concert.

Some portions of the code have been rewritten to reduce difficulty in
testing.

In addition, the predictor_c0 coefficients have been fixed.

+++ opendeplete/results.py
@@ -229,7 +229,7 @@ class Results(object):
 
         handle.create_dataset("seeds", (0, n_stages), maxshape=(None, n_stages), dtype='int64')
 
-        handle.create_dataset("time", (0, 2), maxshape=(None, 2), dtype='int64')
+        handle.create_dataset("time", (0, 2), maxshape=(None, 2), dtype='float64')
 
     def to_hdf5(self, handle, index):
         """ Converts results object into an hdf5 object.
